requirements.txt
Flask
torch
torchvision
Pillow
numpy
opencv-python
gunicorn